% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.R
\name{contract}
\alias{contract}
\title{Lemmatize and stem text}
\usage{
contract(text_vector)
}
\arguments{
\item{text_vector}{Choose character vector or string}
}
\value{
A trimmed character vector or string
}
\description{
This function loads a vector or string and trims the text e.g. hunting = hunt
}
\examples{
contract("hunting")
}
