% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_models.R
\name{download_models}
\alias{download_models}
\title{Clean a text vector}
\usage{
download_models(location)
}
\arguments{
\item{location}{Choose a filepath to save the classification models}
}
\value{
A character vector or string
}
\description{
This function loads a vector os string of type character and tidies the text
}
