% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hun_class.R
\name{hun_class}
\alias{hun_class}
\title{Hunting classifications of tweets}
\usage{
hun_class(type = "full H", directory)
}
\arguments{
\item{type}{Text can be run through five different models: "full H" (default), "relevance", "stance H", "full FH", and "stance FH". See 'Value' for details.}

\item{directory}{Choose directory containing models sourced from download_models function}
}
\value{
A character vector of classifications.



full H classifications (0.86 accuracy): Irrelevant, Against-hunting, Pro-hunting


relevance classifications (0.89 accuracy): Irrelevant, Relevant


stance H classifications (0.99 accuracy): Against-hunting, Pro-hunting


full FH classifications (0.89 accuracy): Irrelevant, Against-hunting, Pro-hunting, Pro-fishing


stance FH classifications (0.97 accuracy): Against-hunting, Pro-hunting, Pro-fishing
}
\description{
This function recevies a vector of tweets, and produces classifications that describe the tweets stance on hunting. Classification accuracy ~85%
}
\examples{
hun_class(type = "full H", directory = "models/classecol-models-master/")
}
