% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_eng.R
\name{lang_eng}
\alias{lang_eng}
\title{Identify if text language is english}
\usage{
lang_eng(text_vector)
}
\arguments{
\item{text_vector}{Choose character vector or string}
}
\value{
A trimmed character vector or string
}
\description{
This function loads a vector or string and checks if it is english
}
\examples{
lang_eng("hunting")
}
