% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_classecol.R
\name{load_classecol}
\alias{load_classecol}
\title{Loads classecol dependencies}
\usage{
load_classecol(download_models = F, download_modules = F, link_py = T)
}
\arguments{
\item{download_models}{Confirm whether the classecol python models should be downloaded (TRUE) or not (FALSE)}

\item{download_modules}{Confirm whether thepython modules should be downloaded (TRUE) or not (FALSE)}

\item{link_py}{Link python to R, select TRUE or FALSE}
}
\value{
No product returned.
}
\description{
This function loads the important dependencies of the classecol pacakge. For classecol's text classification models to work, the text models need to be downloaded (download_models = T), the python modules need to be installed (download_modules = T), and python needs to be linked to R (link_py = T). Once the download_models and download_modules functions have ran once, the information will be stored permanently. However, link_py needs be ran every time a fresh R environment is opened.
}
\examples{
load_classecol(download_models = T, download_modules = T, link_py = T)
}
