% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat_class.R
\name{nat_class}
\alias{nat_class}
\title{Nature classifications of tweets}
\usage{
nat_class(type = "trimmed", directory)
}
\arguments{
\item{type}{Text can be run through four different models: "trimmed"(default), "full", "relevance", and "stance". See 'Value' for details.}

\item{directory}{Choose directory containing models sourced from download_models function}
}
\value{
A character vector of classifications.



trimmed classifications (0.82 accuracy): Irrelevant, Pro-nature positive phrasing, Pro-nature negative phrasing


full classifications (0.81 accuracy): Irrelevant, Pro-nature positive phrasing, Pro-nature negative phrasing, Against-nature


relevance classifications (0.90 accuracy): Irrelevant, Relevant


stance classifications (0.92 accuracy): Pro-nature positive phrasing, Pro-nature negative phrasing
}
\description{
This function recevies a vector of tweets, and produces classifications that describe the tweets stance on nature. Classification accuracy ~80%
}
\examples{
nat_class(type = "trimmed", directory = "models/classecol-models-master/")
}
