% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valence.R
\name{valence}
\alias{valence}
\title{Valence of text}
\usage{
valence(text_vector)
}
\arguments{
\item{text_vector}{Choose character vector or string}
}
\value{
A data frame in the same order as the input, with the following columns: negator, amplifier, de-amplifier, and ad-conjunction. See the lexicon package vignette for details.
}
\description{
This function loads a vector or string and idnetifies valence terms using trinkers lexicon package
}
\examples{
valence("I am not the cat in the hat, I am just a very massive feather")
}
