% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_class.R
\name{bio_class}
\alias{bio_class}
\title{Biographical classifications of twitter users}
\usage{
bio_class(type = "full", directory)
}
\arguments{
\item{type}{Text can be run through three different models: "full"(default), "trimmed person", and "trimmed expert". See 'Value' for details.}

\item{directory}{Choose directory containing models sourced from download_models function}
}
\value{
A character vector of classifications.



full classifications (0.79 accuracy): Expert, Person, Organisations, Nature-organisation


trimmed person classifications (0.87 accuracy): Person, Organisations


trimmed expert (0.84 accuracy): Expert, Person, Organisations
}
\description{
This function recevies a vector of screen names and user bios, and produces classifications that describe the user. Classification accuracy ~80%
}
\examples{
bio_class(type = "full", directory = "models/classecol-models-master/")
}
