% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_class.R
\name{bio_class}
\alias{bio_class}
\title{Biographical classifications of twitter users}
\usage{
bio_class(text_vector, type = "full")
}
\arguments{
\item{text_vector}{Character vector of text data describing the the combination of names and descriptions}

\item{type}{Text can be run through three different models: "full"(default), "person", and "expert". See 'Value' for details.}
}
\value{
A character vector of classifications.



full classifications (0.79 accuracy): Expert, Person, Other, Nature organisation


person classifications (0.87 accuracy): Person, Other


expert classifications (0.94 accuracy): Expert, Person
}
\description{
This function recevies a vector of screen names and user bios, and produces classifications that describe the user. Classification accuracy ~80%
}
\examples{
hun_class(text_vector = df$text, type = "full")
}
