% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_class.R
\name{bio_class}
\alias{bio_class}
\title{Clean a text vector}
\usage{
bio_class(type = "full", directory)
}
\arguments{
\item{type}{Do you want the text to receive a simple or full clean. Default to 'full'}

\item{directory}{Choose cdirectory containing models}
}
\value{
A character vector or string
}
\description{
This function loads a vector os string of type character and tidies the text
}
\examples{
bio_class(type = "full", directory = "models/classecol-models-master/")
}
