% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_class.R
\name{bio_class}
\alias{bio_class}
\title{Biographical classifications of twitter users}
\usage{
bio_class(type = "Full", directory)
}
\arguments{
\item{type}{Text can be run through three different models: "Full"(default), "Person", and "Expert". See 'Value' for details.}

\item{directory}{Choose directory containing models sourced from download_models function}
}
\value{
A character vector of classifications.



Full classifications (0.79 accuracy): Expert, Person, Other, Nature organisation


Person classifications (0.87 accuracy): Person, Other


Expert (0.94 accuracy): Expert, Person
}
\description{
This function recevies a vector of screen names and user bios, and produces classifications that describe the user. Classification accuracy ~80%
}
\examples{
bio_class(type = "Full", directory = "models/classecol-models-master/")
}
