% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean a text vector}
\usage{
clean(
  text_vector,
  hashtag = TRUE,
  abbrev = TRUE,
  emoti = TRUE,
  slang = TRUE,
  grade = TRUE,
  rating = TRUE,
  url = TRUE,
  special_other = TRUE
)
}
\arguments{
\item{infile}{special_other: do you want to remove any remaining special characters, punctions and numbers. Defaults to TRUE}
}
\value{
a character vector
}
\description{
This function loads a text vector
}
