% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hun_class.R
\name{hun_class}
\alias{hun_class}
\title{Hunting classifications of tweets}
\usage{
hun_class(text_vector, type = "full")
}
\arguments{
\item{text_vector}{Character vector of text data}

\item{type}{Text can be run through three different models: "full" (default), "relevance", and "stance". See 'Value' for details.}
}
\value{
A character vector of classifications.


relevance classifications (0.89 accuracy): Irrelevant, Relevant



stance classifications (0.99 accuracy): Against-hunting, Pro-hunting



full classifications (0.87 accuracy): Irrelevant, Against-hunting, Pro-hunting
}
\description{
This function recevies a vector of tweets, and produces classifications that describe the tweets stance on hunting. Classification accuracy ~85%
}
\examples{
hun_class(text_vector = df$text, type = "full")
}
