% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hun_class.R
\name{hun_class}
\alias{hun_class}
\title{Hunting classifications of tweets}
\usage{
hun_class(type = "Full", directory)
}
\arguments{
\item{type}{Text can be run through three different models: "Full" (default), "Relevance", and "Stance". See 'Value' for details.}

\item{directory}{Choose directory containing models sourced from download_models function}
}
\value{
A character vector of classifications.


Relevance classifications (0.89 accuracy): Irrelevant, Relevant



Stance classifications (0.99 accuracy): Against-hunting, Pro-hunting



Full classifications (0.87 accuracy): Irrelevant, Against-hunting, Pro-hunting
}
\description{
This function recevies a vector of tweets, and produces classifications that describe the tweets stance on hunting. Classification accuracy ~85%
}
\examples{
hun_class(type = "Full", directory = "models/classecol-models-master/")
}
